Added cortx-ha sequnce in deploy
